renamed all the .py of the old architecture to be added with old prefix.
